refactor: 분실물 상세 페이지 이미지 컴포넌트 분리
