+++ src/pages/Articles/LostItemDetailPage/LostItemDetailPage.module.scss
@@ -118,34 +118,6 @@
     padding: 0 24px;
   }
 
-  &__images {
-    margin-top: 40px;
-
-    @include media.media-breakpoint(mobile) {
-      margin-top: 20px;
-    }
-  }
-
-  &__navigation {
-    height: 30px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 20px;
-    box-sizing: border-box;
-    gap: 8px;
-
-    & button {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-
-    @include media.media-breakpoint(mobile) {
-      margin-top: 16px;
-    }
-  }
-
   &__content {
     width: 794px;
     margin-top: 64px;
@@ -236,56 +208,6 @@
         width: 16px;
         height: 16px;
       }
-    }
-  }
-}
-
-.images {
-  position: relative;
-  width: 586px;
-  height: 527px;
-
-  @include media.media-breakpoint(mobile) {
-    width: 100%;
-    height: auto;
-    padding: 0 24px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  &__image {
-    width: 586px;
-    height: 527px;
-    object-fit: contain;
-
-    @include media.media-breakpoint(mobile) {
-      width: 100%;
-      height: auto;
-    }
-  }
-
-  &__button-left {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-
-  &__button {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-
-    &--left {
-      left: 17px;
-    }
-
-    &--right {
-      right: 17px;
-    }
-
-    &--hidden {
-      display: none;
     }
   }
 }
