+++ src/pages/Articles/LostItemDetailPage/components/DeleteModal/DeleteModal.module.scss
